Build and push a docker image checks node security
Builds a docker image that checks an auth0 service
node version security. The Jenkins job that builds
and publishes the docker image is triggered each time
the auth0-configuration master branch changes. It copies
the versions.json file the auth0-configuration archieved
and bundles it within the docker image. The versions.json
contains a mapping between services and node versions.